+++ changelog.md
@@ -64,3 +64,5 @@
 ## 0.3.12
 - Now shows if asset is already in an other kit
 - Added a character limit to the asset name and team name for the 29x90mm label
+- Added status to asset and kit
+- Updated logo and favicon